requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
PyMuPDF
langchain
langchain-openai
